Fix code review issues: type consistency, cycling logic, heading loop, constant placement
Agent-Logs-Url: https://github.com/HechaoYannet/readwiseAI/sessions/4a5f1bd0-c5f3-4949-ad34-32d0c57cb3cd

+++ app/models/forgetting.py
@@ -58,7 +58,7 @@ class SM2Item(BaseModel):
             elif self.repetitions == 1:
                 self.interval_days = 6
             else:
-                self.interval_days = round(self.interval_days * self.easiness)
+                self.interval_days = int(round(self.interval_days * self.easiness))
             self.repetitions += 1
         else:
             # Reset on failure
